requirements.txt
streamlit>=1.30.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
cvxpy>=1.4.0
statsmodels>=0.14.0
plotly>=5.18.0
pyyaml>=6.0
matplotlib>=3.7.0
arch>=6.0.0
requests>=2.31.0
lxml>=5.0.0
